+++ api/src/api.rs
@@ -48,7 +48,7 @@ use crate::dynamic_settings;
 use crate::handlers::machine_validation::{
     add_update_machine_validation_external_config, get_machine_validation_external_config,
     get_machine_validation_results, get_machine_validation_runs, mark_machine_validation_complete,
-    persist_validation_result,
+    on_demand_machine_validation, persist_validation_result,
 };
 use crate::ib::{IBFabricManager, DEFAULT_IB_FABRIC_NAME};
 use crate::logging::log_limiter::LogLimiter;
@@ -4013,6 +4013,13 @@ impl Forge for Api {
         .await?;
 
         Ok(Response::new(rpc::AdminPowerControlResponse {}))
+    }
+
+    async fn on_demand_machine_validation(
+        &self,
+        request: tonic::Request<rpc::MachineValidationOnDemandRequest>,
+    ) -> Result<tonic::Response<rpc::MachineValidationOnDemandResponse>, Status> {
+        on_demand_machine_validation(self, request).await
     }
 }