+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   ```
 
 ### Changed
+- FORGE-4465 - On-demand machine-validation API
 ### Fixed
 - Fixes redirect after Power Actions or BMC Reset have been issued on the carbide admin web UI.